alx_travel_app/requirement.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
django-environ==0.11.2
drf-yasg==1.21.7
mysqlclient==2.2.0
redis==5.0.1
kombu==5.3.4
billiard==4.2.0
